+++ ecom-backend/src/main/java/com/ecommerce/project/service/order/event/OrderPlacedEvent.java
@@ -2,6 +2,8 @@ package com.ecommerce.project.service.order.event;
 
 import com.ecommerce.project.payload.OrderDTO;
 
+import java.math.BigDecimal;
+
 /**
  * Published after an order is committed (customer checkout or guest checkout).
  * <p>
@@ -9,5 +11,5 @@ import com.ecommerce.project.payload.OrderDTO;
  * so listeners can run on another thread without hitting
  * {@code LazyInitializationException}.
  */
-public record OrderPlacedEvent(String email, Long orderId, double totalAmount, OrderDTO orderDTO) {
+public record OrderPlacedEvent(String email, Long orderId, BigDecimal totalAmount, OrderDTO orderDTO) {
 }
